Extracted LogJsonOutputUtils. Fixed stylecop warnings
